增加了 personalAciton, 但是还没有测试

+++ src/com/istc/Service/EntityService/IntervieweeService.java
@@ -23,6 +23,8 @@ public class IntervieweeService {
     private transient MemberDAO memberDAO;
 
     public boolean exist(String id){
+        if (id==null)
+            return false;
         return intervieweeDAO.exist(id);
     }
 
@@ -30,23 +32,26 @@ public class IntervieweeService {
         return intervieweeDAO.findAll();
     }
 
-    public void setIntervieweesToMembers(String[] intervieweeIDs){
-        List<Interviewee> list = intervieweeDAO.get(intervieweeIDs);
-        intervieweeDAO.delete(intervieweeIDs);
-        ClassTypeConverter converter = ClassTypeConverter.getInstance();
-        Member[] members = null;
-        try {
-            members = (Member[]) converter.convert(list, Member.class);
-        } catch (Exception e) {
-            System.out.println("IntervieweeService.java 文件中的成员类型转换失败\n" +
-                    "设定converter设定阈值时访问异常");
-            e.printStackTrace();
-            return;
+    public void setIntervieweesToMembers(String[] intervieweeIDs) {
+        if (intervieweeIDs[0] != null) {
+            List<Interviewee> list = intervieweeDAO.get(intervieweeIDs);
+            intervieweeDAO.delete(intervieweeIDs);
+            ClassTypeConverter converter = ClassTypeConverter.getInstance();
+            Member[] members = null;
+            try {
+                members = (Member[]) converter.convert(list, Member.class);
+            } catch (Exception e) {
+                System.out.println("IntervieweeService.java 文件中的成员类型转换失败\n" +
+                        "设定converter设定阈值时访问异常");
+                e.printStackTrace();
+                return;
+            }
+            memberDAO.save(members);
         }
-        memberDAO.save(members);
     }
-
+//即使是申请面试也需要基本id（primarily key）
     public void add(Interviewee interviewee){
+        if (interviewee!=null&&interviewee.getID()!=null)
         intervieweeDAO.save(interviewee);
     }
 }
